create eslint plugin to catch optional tool inputs (#19)
There is this annoying problem that models like GPT-5 do not support
tools with any optional parameters. Declaring a parameter as not
required in the JSON schema will effectively make the tool unusable with
such models.

Since this is somewhat unexpected, and people/agents naturally tend to
design APIs with optional arguments, this keeps coming up. Luckily, we
have caught many of these in PR reviews so far, but we can do better
with some automation.

This implements an eslint rule to do exactly that. Paired with a GitHub
workflow to run eslint, this will guard against PRs with this mistake.

<img width="1401" height="247" alt="image"
src="https://github.com/user-attachments/assets/872236ee-603d-4322-b4bf-68547b2e6b15"
/>